Moved client to client module
